Move service.ts to private
