+++ packages/app/src/cli/models/extensions/ui.ts
@@ -135,7 +135,7 @@ export class UIExtensionInstance<TConfiguration extends BaseConfigContents = Bas
   }
 
   async publishURL(options: {orgId: string; appId: string; extensionId?: string}) {
-    const partnersFqdn = await environment.fqdn.partners()
+    const partnersFqdn = await environment.fqdn.partnersFqdn()
     const parnersPath = this.specification.partnersWebIdentifier
     return `https://${partnersFqdn}/${options.orgId}/apps/${options.appId}/extensions/${parnersPath}/${options.extensionId}`
   }
