+++ packages/app/src/cli/services/dev/select-store.test.ts
@@ -6,6 +6,7 @@ import {beforeEach, describe, expect, it, vi} from 'vitest'
 import {environment} from '@shopify/cli-kit'
 import {partnersRequest} from '@shopify/cli-kit/node/api/partners'
 import {ensureAuthenticatedPartners} from '@shopify/cli-kit/node/session'
+import {isSpinEnvironment} from '@shopify/cli-kit/node/environment/spin'
 
 const ORG1: Organization = {id: '1', businessName: 'org1', appsNext: true}
 const STORE1: OrganizationStore = {
@@ -42,6 +43,8 @@ beforeEach(() => {
   vi.mock('@shopify/cli-kit/node/api/partners')
   vi.mock('@shopify/cli-kit/node/session')
   vi.mocked(ensureAuthenticatedPartners).mockResolvedValue('token')
+  vi.mock('@shopify/cli-kit/node/environment/spin')
+  vi.mocked(isSpinEnvironment).mockReturnValue(false)
   vi.mock('@shopify/cli-kit', async () => {
     const cliKit: any = await vi.importActual('@shopify/cli-kit')
     return {
@@ -50,9 +53,6 @@ beforeEach(() => {
         fetch: vi.fn(),
       },
       environment: {
-        service: {
-          isSpinEnvironment: vi.fn(),
-        },
         local: {
           firstPartyDev: vi.fn(),
           isUnitTest: vi.fn(() => true),
@@ -94,7 +94,7 @@ describe('selectStore', async () => {
   it('not prompts user to convert store to non-transferable if selection is invalid inside spin instance and first party', async () => {
     // Given
     vi.mocked(selectStorePrompt).mockResolvedValueOnce(STORE2)
-    vi.mocked(environment.service.isSpinEnvironment).mockReturnValue(true)
+    vi.mocked(isSpinEnvironment).mockReturnValue(true)
     vi.mocked(environment.local.firstPartyDev).mockReturnValue(true)
 
     // When
