+++ packages/app/src/cli/services/dev/select-store.ts
@@ -9,9 +9,10 @@ import {
 import {error, output, ui, environment} from '@shopify/cli-kit'
 import {partnersRequest} from '@shopify/cli-kit/node/api/partners'
 import {sleep} from '@shopify/cli-kit/node/system'
+import {isSpinEnvironment} from '@shopify/cli-kit/node/environment/spin'
 
 const CreateStoreLink = async (orgId: string) => {
-  const url = `https://${await environment.fqdn.partners()}/${orgId}/stores/new?store_type=dev_store`
+  const url = `https://${await environment.fqdn.partnersFqdn()}/${orgId}/stores/new?store_type=dev_store`
   return (
     `Looks like you don't have a dev store in the Partners org you selected. ` +
     `Keep going — create a dev store on Shopify Partners:\n${url}\n`
@@ -105,7 +106,7 @@ export async function convertToTestStoreIfNeeded(
   /**
    * Is not possible to convert stores to dev ones in spin environmets. Should be created directly as development.
    */
-  if (environment.service.isSpinEnvironment() && environment.local.firstPartyDev()) return
+  if (isSpinEnvironment() && environment.local.firstPartyDev()) return
   if (!store.transferDisabled && !store.convertableToPartnerTest) {
     throw new error.Abort(
       `The store you specified (${store.shopDomain}) is not a dev store`,
