+++ packages/cli-kit/src/environment.ts
@@ -1,5 +1,4 @@
 import * as local from './environment/local.js'
-import * as service from './environment/service.js'
 import * as fqdn from './environment/fqdn.js'
 
-export {local, service, fqdn}
+export {local, fqdn}
