+++ packages/cli-kit/src/private/node/session/authorize.test.ts
@@ -3,7 +3,7 @@ import {clientId} from './identity.js'
 import {listenRedirect} from './redirect-listener.js'
 import {randomHex, base64URLEncode} from '../../../public/node/crypto.js'
 import {openURL} from '../../../public/node/system.js'
-import {identity} from '../../../environment/fqdn.js'
+import {identityFqdn} from '../../../environment/fqdn.js'
 import {terminateBlockingPortProcessPrompt} from '../../../ui.js'
 import {CancelExecution} from '../../../error.js'
 import {checkPort} from 'get-port-please'
@@ -35,7 +35,7 @@ describe('authorize', () => {
     vi.mocked(base64URLEncode).mockReturnValueOnce('verifier')
     vi.mocked(base64URLEncode).mockReturnValueOnce('challenge')
     vi.mocked(listenRedirect).mockResolvedValue({code: 'code', state: 'state'})
-    vi.mocked(identity).mockResolvedValue('fqdn.com')
+    vi.mocked(identityFqdn).mockResolvedValue('fqdn.com')
     vi.mocked(clientId).mockReturnValue('clientId')
 
     // When
@@ -86,7 +86,7 @@ describe('authorize', () => {
     vi.mocked(base64URLEncode).mockReturnValueOnce('verifier')
     vi.mocked(base64URLEncode).mockReturnValueOnce('challenge')
     vi.mocked(listenRedirect).mockResolvedValue({code: 'code', state: 'state'})
-    vi.mocked(identity).mockResolvedValue('fqdn.com')
+    vi.mocked(identityFqdn).mockResolvedValue('fqdn.com')
     vi.mocked(clientId).mockReturnValue('clientId')
     vi.mocked(checkPort).mockResolvedValue(false)
     vi.mocked(terminateBlockingPortProcessPrompt).mockResolvedValue(true)
