+++ packages/cli-kit/src/private/node/session/authorize.ts
@@ -3,7 +3,7 @@ import {listenRedirect} from './redirect-listener.js'
 import {base64URLEncode, randomBytes, randomHex, sha256} from '../../../public/node/crypto.js'
 import {openURL} from '../../../public/node/system.js'
 import {Abort, CancelExecution} from '../../../error.js'
-import {identity as identityFqdn} from '../../../environment/fqdn.js'
+import {identityFqdn} from '../../../environment/fqdn.js'
 import * as output from '../../../output.js'
 import {keypress, terminateBlockingPortProcessPrompt} from '../../../ui.js'
 import {checkPort as isPortAvailable} from 'get-port-please'
